+++ common/src/main/java/net/drgmes/dwm/network/server/TardisConsoleUnitMonitorWaypointDeletePacket.java
@@ -39,6 +39,11 @@ public record TardisConsoleUnitMonitorWaypointDeletePacket(
             ServerWorld tardisWorld = DimensionHelper.getModWorld(payload.tardisId, player.getServer());
 
             TardisStateManager.get(tardisWorld).ifPresent((tardis) -> {
+                if (!tardis.checkAccess(player, true, false)) {
+                    player.sendMessage(DWM.TEXTS.TARDIS_NOT_ALLOWED, true);
+                    return;
+                }
+
                 boolean flag = tardis.getSystem(TardisSystemFlight.class).deleteWaypointEntry(payload.waypointEntry);
                 if (flag) player.sendMessage(DWM.TEXTS.MONITOR_WAYPOINT_REMOVED, true);
             });
